requirements.txt
aiofiles==24.1.0
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
annotated-types==0.7.0
anyio==4.11.0
archspec @ file:///croot/archspec_1709217642129/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
attrs==25.4.0
beautifulsoup4==4.12.3
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
Brotli==1.1.0
certifi==2025.10.5
cffi @ file:///C:/b/abs_78eb1_vq6z/croot/cffi_1714483206096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
choreographer==1.1.2
click==8.3.0
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
conda @ file:///D:/bld/conda_1736460255649/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///C:/b/abs_831_pypuuz/croot/conda-package-handling_1718138287790/work
conda_package_streaming @ file:///C:/b/abs_fe6zb0_j8_/croot/conda-package-streaming_1718136098642/work
contourpy==1.3.3
cryptography @ file:///C:/b/abs_97lgskwerx/croot/cryptography_1714664736250/work
cycler==0.12.1
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1758162063/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
duckdb==1.4.1
EasyProcess==1.1
entrypoint2==1.1
et_xmlfile==2.0.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.119.0
ffmpy==0.6.3
filelock==3.20.0
fonttools==4.60.1
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
fsspec==2025.9.0
google-api-core==2.30.1
google-auth==2.49.1
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1762350976/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.2
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
kaleido==1.0.0
kiwisolver==1.4.9
libmambapy @ file:///C:/b/abs_89suv9zwkz/croot/mamba-split_1714484512428/work/libmambapy
logistro==2.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
menuinst @ file:///C:/b/abs_e7zsrwvoy0/croot/menuinst_1718132546150/work
mysql-connector-python==8.3.0
narwhals==2.8.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
nltk==3.9.1
numpy==2.1.2
openpyxl==3.1.5
orjson==3.11.3
outcome==1.3.0.post0
packaging==26.0
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
patool==1.12
patsy==1.0.1
pillow==11.3.0
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
plotly==6.3.1
pluggy @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pluggy_1699472504117/work
polars==1.34.0
polars-runtime-32==1.34.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.11
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==16.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.5
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487136/work
PyYAML==6.0.1
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387025/work
RapidFuzz==3.14.1
rarfile==4.1
regex==2025.11.3
requests==2.32.3
rich==14.2.0
ruamel.yaml @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/ruamel.yaml_1699483184324/work
ruff==0.14.1
safehttpx==0.1.6
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.24.0
semantic-version==2.10.0
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.20.2
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.48.0
statsmodels==0.14.5
threadpoolctl==3.6.0
tomlkit==0.13.3
tornado @ file:///D:/bld/tornado_1762506742870/work
tqdm==4.66.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
trio==0.31.0
trio-websocket==0.12.2
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typer==0.19.2
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.5.0
uvicorn==0.38.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wheel==0.43.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wsproto==1.2.0
xlsxwriter==3.2.9
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard @ file:///C:/b/abs_e46tf0_91g/croot/zstandard_1714678303615/work
